Add simple warrior display into round results
